# swiss: Go implementation of Swiss Tables