--- a/README.md
+++ b/README.md
@@ -1,0 +1,6 @@
+# Swiss Map
+
+`swiss.Map` is a Go implementation of [Google's Swiss Tables hash table
+design](https://abseil.io/about/design/swisstables). The [Rust version of
+Swiss Tables](https://github.com/rust-lang/hashbrown) is now the `HashMap`
+implementation in the Rust standard library.