… go.sum
… map.go