@petermattis 2024-02-09
This is pretty much exactly https://github.com/cockroachdb/cockroach/pull/118742. Rather than fleshing out all of the TODOs and having you re-review the code again, I felt it might be easier to review this as-is, get it checked in and follow up with subsequent PRs to address the TODOs and set up CI.